+++ RSPiX/Src/GREEN/BLiT/_ic.h
@@ -58,7 +58,7 @@ inline u16Frac* u16fStrafe256(USHORT usNum,USHORT usDen)
 	u16Frac u16fInc;
 	MakeProper(u16fInc,usNum,usDen); // the 2 part delta
 
-	ULONG ulNumInc = 0;
+	uint32_t ulNumInc = 0;
 	for (int16_t i = 1; i < 256 ; i++)
 		{
 		pu16fNew[i].delta = pu16fNew[i-1].delta + u16fInc.delta;
